cdk/requirements.txt
aws-cdk-lib<3.0.0,>=2.114.1
constructs>=10.0.0,<11.0.0
